Refactored file structure to be simpler
